+++ src/main/java/io/nem/sdk/dto/BlockchainStorageInfoDTO.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.nem.sdk.model.blockchain;
+package io.nem.sdk.dto;
 
 import com.google.gson.annotations.SerializedName;
 import io.swagger.annotations.ApiModelProperty;
