+++ src/main/java/io/nem/sdk/dto/MosaicProperties.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2018 NEM
+ * Copyright 2019 NEM
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.nem.sdk.model.mosaic;
+package io.nem.sdk.dto;
 
 import org.apache.commons.lang3.Validate;
