+++ src/main/java/io/nem/sdk/dto/NamespaceDTO.java
@@ -14,14 +14,12 @@
  * limitations under the License.
  */
 
-package io.nem.sdk.model.namespace;
+package io.nem.sdk.dto;
 
 import com.google.gson.annotations.SerializedName;
-import io.nem.sdk.model.account.AliasDTO;
 import io.nem.sdk.model.transaction.UInt64;
 import io.swagger.annotations.ApiModelProperty;
 
-import java.math.BigDecimal;
 import java.util.Objects;
 
 /**
