+++ src/main/java/io/nem/sdk/dto/NamespaceMosaicMetaDTO.java
@@ -14,18 +14,12 @@
  * limitations under the License.
  */
 
+package io.nem.sdk.dto;
 
-package io.nem.sdk.model.namespace;
+import com.google.gson.annotations.SerializedName;
+import io.swagger.annotations.ApiModelProperty;
 
 import java.util.Objects;
-import com.google.gson.TypeAdapter;
-import com.google.gson.annotations.JsonAdapter;
-import com.google.gson.annotations.SerializedName;
-import com.google.gson.stream.JsonReader;
-import com.google.gson.stream.JsonWriter;
-import io.swagger.annotations.ApiModel;
-import io.swagger.annotations.ApiModelProperty;
-import java.io.IOException;
 
 /**
  * NamespaceMosaicMetaDTO
